Revision for EMF Compatibility
In attempting to have a well defined process and compatible models I did
the following;

• Made model changes noted below
• Ran scripts “ownedEnds” and “MOFIze”
• Made “definedIn” non-navigable to removed opposite end.
• Saved as “SMIFMOFModel”
• Exported as “.ecore”, .emof file and UML 2.5 XMI and Eclipse UML file.
Let me know which one you use.
The above is checked in under “NextSubmission”  - the EMF files are
under “Non Normative”
Thanks!

-Cory
From: David Zenie [mailto:davidzenie@gmail.com]
Sent: Saturday, February 4, 2017 3:27 PM
To: cory@casanave.us; 'Jim Logan' <jlogan@nomagic.com>; 'Nilesh Kawane'
<nkawane@nomagic.com>
Subject: RE: Working with Dave's SMIF EMF Model

I can help with the first part (am sending the .ecore used (this should
be the same as the one I sent in the zip with the generated projects
except it also fixes Package names to please Ecore)).  Next will be
giving it to MD to import with a list of what was changed from the
original uml2ecore so it would validate and generate in EMF – it is FYI
in case there are problems mapping it back into uml from the ecore
attached.

Most of the fixes below have been in other emails, this list puts them
all together:

Ecore validation’s found 7 container problems after importing the
original *OwnedEnds.uml into a new EMF Project.  Summary of the ecore
changes:

1. Imported ecore changed  to not disturb the foundational containment,
defines:Thing – removed the other end (the “opposite” in ecore)
definedIn:LexicalScope which then fixes 5 of the 7 containment
validation errors.  Information lost doing this is
definedIn:LexicalScope, an obvious opposite of defines:Thing, and the
UML {subsets inContextOf}.  It works so far for ecore, don’t know about
UML.
Will to as a manual step to produce the MOF model
2. The other 2 containment errors are fixed in Pattern,
ownsVariable:Pattern Containment set to False.  However, not sure this
ecore fix is consistent with the model’s intent.
Removed these – could “shared” containment be OK with EMF??
3. And the duplicate property name is fixed – in ScalarQuantity
hasValue is changed to hasNumberValue.  Again just an ecore fix, there
may be better ways.
Made blank name so it will just be a constraint
4. Other changes in the ecore are in Values, datatypes RealNumber and
Text have their InstanceTypeName property set to Real and String.  This
should have been done by the EMF UML import but expect it doesn’t
understand Dependencys.
I changed these to the EMF values – would prefer not to have to do this.
I don’t know how to set an “InstanceTypeName” – do you?
5. A concrete class named Repository -> Namespace contains abstract
class LexicalScope, as an EMF Root container must be concrete.
Made it concrete – but that is not the intent.
6. All package names were changed to lower case and simplified because
of EMF generation problems.  They can be set back to their intended
values and if there are problems generating later we can identify and
fix them individually – it could be a result of too many mappings when I
was exploring how to start.
Didn’t make this change.

Also – changed “Rules” to “Constraints” as that is what is really
defined in that package.